requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
pandas==2.1.3
openpyxl==3.1.2
pyodbc==5.0.1
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0